+++ compiler-core/src/type_/error.rs
@@ -720,6 +720,14 @@ pub enum Error {
     ///
     ListPrependWithoutElements {
         location: SrcSpan,
+        /// We hold onto the location of the tail expression for the code action.
+        ///
+        /// ```gleam
+        /// const wibble = [..wobble]
+        ///                   ^^^^^^ here
+        /// ```
+        ///
+        tail_location: SrcSpan,
     },
 
     /// When a constant contains a todo.
@@ -1416,7 +1424,10 @@ impl Error {
             | Error::QualifiedTypeMissingName { location }
             | Error::TodoConstant { location }
             | Error::LowercaseBoolPattern { location }
-            | Error::ListPrependWithoutElements { location } => location.start,
+            | Error::ListPrependWithoutElements {
+                location,
+                tail_location: _,
+            } => location.start,
             Error::UnknownLabels { unknown, .. } => {
                 unknown.iter().map(|(_, s)| s.start).min().unwrap_or(0)
             }